requirements.txt
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-customvision
azure-cognitiveservices-vision-face
azure-storage-blob
fastapi
gunicorn
mysql-connector-python
numpy
oauthlib
pillow
wheel
pymongo
PyPDF2
python-dotenv
python-multipart
scikit-learn
nltk
openai
pandas